+++ packages/spotlight/Spottable/Spottable.js
@@ -9,6 +9,7 @@
 import {forward} from '@enact/core/handle';
 import hoc from '@enact/core/hoc';
 import {is} from '@enact/core/keymap';
+import pick from 'ramda/src/pick';
 import React from 'react';
 
 import Spotlight from '../src/spotlight';
@@ -62,6 +63,13 @@ const forwardEnter = (keyEvent, mouseEvent) => (props) => {
 	};
 };
 
+const keyDownEventProps = ['altKey', 'ctrlKey', 'currentTarget', 'detail', 'key', 'keyCode',
+	'metaKey', 'shiftKey', 'target', 'type', 'which'];
+
+const makeEvent = (type, ev) => {
+	return {...ev, type};
+};
+
 /**
  * Default configuration for Spottable
  *
@@ -78,7 +86,18 @@ const defaultConfig = {
 	 * @public
 	 * @memberof spotlight/Spottable.Spottable.defaultConfig
 	 */
-	emulateMouse: true
+	emulateMouse: true,
+
+	/**
+	 * Number of milliseconds to wait before forwarding a `keyup` event, due to
+	 * focus changing from an unreleased `keydown` event.
+	 *
+	 * @type {Number}
+	 * @default 0
+	 * @public
+	 * @memberof spotlight/Spottable.Spottable.defaultConfig
+	 */
+	selectionKeyUpDelay: 0
 };
 
 /**
@@ -96,13 +115,16 @@ const defaultConfig = {
  * @returns {Function} Spottable
  */
 const Spottable = hoc(defaultConfig, (config, Wrapped) => {
-	const {emulateMouse} = config;
+	const {emulateMouse, selectionKeyUpDelay} = config;
 	const forwardBlur = forward('onBlur');
 	const forwardFocus = forward('onFocus');
 	const forwardEnterKeyPress = forwardEnter('onKeyPress', 'onClick');
 	const forwardEnterKeyDown = forwardEnter('onKeyDown', 'onMouseDown');
 	const forwardEnterKeyUp = forwardEnter('onKeyUp', 'onMouseUp');
 	const forwardKeyDown = forward('onKeyDown');
+	const forwardKeyUp = forward('onKeyUp');
+	const forwardClick = forward('onClick');
+	const forwardMouseDown = forward('onMouseDown');
 
 	return class extends React.Component {
 		static displayName = 'Spottable'
@@ -183,9 +205,26 @@ const Spottable = hoc(defaultConfig, (config, Wrapped) => {
 
 		constructor (props) {
 			super(props);
+			this.enterKeyDownEvent = null;
 			this.state = {
 				spotted: false
 			};
+		}
+
+		componentDidUpdate (prevProps, prevState) {
+			// if a blur occured during a enter-key press/hold, as a result of a programmatic focus change
+			if (prevState.spotted && !this.state.spotted && this.enterKeyDownEvent) {
+				// Certain components require time to perform an action (animation via an applied class) due to
+				// forwarding a `keydown` event, while being selected. Immediately calling `forwardEnterKeyUp`
+				// will result in the components being updated too quickly for the animation to begin, providing
+				// no visual feedback of the selection. We perform this behavior in the `componentDidUpdate`
+				// life-cycle instead of immediately within the `onBlur` event to ensure the component tree has
+				// been updated, giving us the full `selectionKeyUpDelay` time to work with.
+				setTimeout(() => {
+					forwardEnterKeyUp(this.props)(makeEvent('onKeyUp', this.enterKeyDownEvent));
+					this.enterKeyDownEvent = null;
+				}, selectionKeyUpDelay);
+			}
 		}
 
 		componentWillUnmount () {
@@ -235,9 +274,37 @@ const Spottable = hoc(defaultConfig, (config, Wrapped) => {
 			}
 
 			if (emulateMouse && !(this.state.spotted && disabled) && is('enter', keyCode)) {
+				this.enterKeyDownEvent = pick(keyDownEventProps, e);
 				forwardEnterKeyDown(this.props)(e);
 			} else {
 				forwardKeyDown(e, this.props);
+			}
+		}
+
+		onKeyUp = (e) => {
+			if (emulateMouse && !(this.state.spotted && this.props.disabled) && is('enter', e.keyCode)) {
+				this.enterKeyDownEvent = null;
+				forwardEnterKeyUp(this.props)(e);
+			} else {
+				forwardKeyUp(e, this.props);
+			}
+		}
+
+		onMouseDown = (e) => {
+			if (!this.state.spotted) {
+				// prevent component from gaining focus
+				e.preventDefault();
+			} else {
+				forwardMouseDown(e, this.props);
+			}
+		}
+
+		onClick = (e) => {
+			// valid (non-disabled) spottable components can be prevented from gaining focus during `mousedown`, specifically
+			// in cases where they exist outside of an active container using `spotlightRestrict: 'self-only'`, so we guard
+			// against forwarding unwanted `click` events
+			if (this.state.spotted) {
+				forwardClick(e, this.props);
 			}
 		}
 
@@ -263,10 +330,12 @@ const Spottable = hoc(defaultConfig, (config, Wrapped) => {
 				rest['onBlur'] = this.onBlur;
 				rest['onFocus'] = this.onFocus;
 				rest['onKeyDown'] = this.onKeyDown;
+				rest['onKeyUp'] = this.onKeyUp;
+				rest['onClick'] = this.onClick;
+				rest['onMouseDown'] = this.onMouseDown;
 
 				if (emulateMouse && !spottableDisabled) {
 					rest['onKeyPress'] = forwardEnterKeyPress(this.props);
-					rest['onKeyUp'] = forwardEnterKeyUp(this.props);
 				}
 				if (rest.className) {
 					rest.className += ' ' + classes;