+++ packages/spotlight/CHANGELOG.md
@@ -6,7 +6,9 @@ The following is a curated list of changes in the Enact spotlight module, newest
 
 ### Fixed
 
-`spotlight.Spotlight` behavior to properly save the last-focused element for nested containers
+- `spotlight/Spotlight` setting the correct active container when calling `focus(containerId)`
+- `spotlight/Spotlight` behavior to follow the various container config rules when navigating between nested containers
+- `spotlight/Spotlight` behavior to properly save the last-focused element for nested containers
 
 ### Removed